[RFE] [WIP] Switch to Python3

+++ foobnix/util/time_utils.py
@@ -4,7 +4,6 @@ Created on Feb 26, 2010
 @author: ivan
 '''
 
-
 def size2text(size):
     if size > 1024 * 1024 * 1024:
         return "%.2f Gb" % (size / (1024 * 1024 * 1024.0))
@@ -17,13 +16,13 @@ def size2text(size):
 def convert_seconds_to_text(time_sec):
         time_sec = int(time_sec)
 
-        hours = time_sec / (60 * 60)
+        hours = int(time_sec / (60 * 60))
         time_sec = time_sec - (hours * 60 * 60)
 
-        mins = time_sec / 60
+        mins = int(time_sec / 60)
         time_sec = time_sec - (mins * 60)
 
-        secs = time_sec
+        secs = int(time_sec)
         if hours > 0:
             return '%(hours)d:%(mins)02d:%(secs)02d' % {'hours' : hours, 'mins': mins, 'secs': secs }
         else:
